Restrict queries by current user

+++ class-19/demo/todo-app/server.js
@@ -95,10 +95,11 @@ function getTasks(request, response) {
   const SQL = `
     SELECT *
     FROM Tasks
+    WHERE user_id = $2
     ORDER BY $1 ASC;
   `;
 
-  client.query(SQL, [sort_by || 'due'])
+  client.query(SQL, [sort_by || 'due', request.user.id])
     .then(results => {
       const { rows } = results;
 
@@ -119,10 +120,11 @@ function getOneTask(request, response) {
     SELECT *
     FROM Tasks
     WHERE id = $1
+    AND user_id = $2
     LIMIT 1;
   `;
 
-  client.query(SQL, [task_id])
+  client.query(SQL, [task_id, request.user.id])
     .then(results => {
       const { rows } = results;
 
@@ -146,11 +148,11 @@ function addTask(request, response) {
   const { title, description, category, contact, status } = request.body;
 
   const SQL = `
-    INSERT INTO tasks (title, description, category, contact, status)
-    VALUES ($1, $2, $3, $4, $5)
+    INSERT INTO tasks (title, description, category, contact, status, user_id)
+    VALUES ($1, $2, $3, $4, $5, $6)
     RETURNING Id
   `;
-  const values = [title, description, category, contact, status];
+  const values = [title, description, category, contact, status, request.user.id];
 
   // POST - REDIRECT - GET
   client.query(SQL, values)
@@ -166,8 +168,9 @@ function deleteOneTask(request, response) {
   const SQL = `
     DELETE FROM Tasks
     WHERE Id = $1
+    AND user_id = $2
   `
-  client.query(SQL, [request.params.task_id])
+  client.query(SQL, [request.params.task_id, request.user.id])
     .then(() => {
       response.redirect('/');
     })
@@ -180,8 +183,9 @@ function editOneTask(request, response) {
     SELECT *
     FROM Tasks
     WHERE Id = $1
+    AND user_id = $2
   `;
-  client.query(SQL, [task_id])
+  client.query(SQL, [task_id, request.user.id])
     .then(results => {
       const task = results.rows[0];
       const viewModel = {
@@ -203,8 +207,9 @@ function updateOneTask(request, response, next) {
       Contact = $4,
       Status = $5
     WHERE Id = $6
+    AND user_id = $7
   `;
-  const parameters = [title, description, category, contact, status, task_id];
+  const parameters = [title, description, category, contact, status, task_id, request.user.id];
   client.query(SQL, parameters)
     .then(() => {
       response.redirect(`/tasks/${task_id}`);
